requirements.txt

python-dotenv
python-telegram-bot==13.15
websockets
pandas
numpy
requests
matplotlib
pillow
